very very basic apikey auth... pls dont hack me...

+++ pages/api/post/delete.ts
@@ -1,24 +1,32 @@
-import operand from 'lib/operand';
-import prisma from '../../../lib/prisma';
+import operand from "lib/operand";
+import prisma from "../../../lib/prisma";
 
 export default async function handler(req, res) {
-  const id = req.body.id
-  console.log(`[delete] received id: ${id}`)
-  const post = await prisma.post.findUnique({
-    where: {
-      id: id
-    }
-  });
-  if (post) {
-    if (post.operandId) {
-      await operand.deleteGroup(post.operandId);
-      console.log(`[operand] deleted group: ${post.operandId}`);
-    }
-    await prisma.post.delete({
+  const apiKey = req.headers["authorization"].split(" ")[1];
+  if (apiKey != process.env.API_KEY) {
+    res.status(401);
+  } else {
+    const id = req.body.id;
+    console.log(`[delete] received id: ${id}`);
+    const post = await prisma.post.findUnique({
       where: {
-        id: id 
+        id: id,
+      },
+    });
+    if (post) {
+      if (post.operandId) {
+        await operand.deleteGroup(post.operandId);
+        console.log(`[operand] deleted group: ${post.operandId}`);
       }
-    })
+      await prisma.post.delete({
+        where: {
+          id: id,
+        },
+      });
+      res.status(200).json({ id: id });
+    } else {
+      res.status(404);
+    }
   }
-  res.status(200).json({ id: id })
+  res.end();
 }
